# docs/swagger.yaml
basePath: /
definitions:
...
  p2-ip-kartikajls_internal_dto.CreateRentalOrderRequest:
    properties:
      camera_id:
        type: integer
    required:
    - camera_id
    type: object
...
  p2-ip-kartikajls_internal_dto.UpdateRentalOrderStatusRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
  p2-ip-kartikajls_internal_helper.Response:
    properties:
      responseCode:
        type: integer
      responseData: {}
      responseMessage:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API untuk aplikasi Rent Camera.
    API ini menyediakan fitur register, login, camera, top up, rental order, dan payment.
    Admin dapat melakukan approval terhadap top up, order, dan payment.
  title: Jati Rent Camera API
  version: "1.0"
paths:
  /admin/orders:
    get:
      description: Admin melihat seluruh rental order
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Get All Orders
      tags:
      - Admin - Order Approve
  /admin/orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Admin menyetujui atau mengubah status rental order user
      parameters:
      - description: Rental Order ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Order Status Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/p2-ip-kartikajls_internal_dto.UpdateRentalOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Approve Rental Order
      tags:
      - Admin - Order Approve
...
  /users/cameras:
    get:
      consumes:
      - application/json
      description: User melihat daftar kamera yang tersedia untuk rental
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Get All Cameras
      tags:
      - User - Camera
  /users/cameras/{id}:
    get:
      consumes:
      - application/json
      description: User melihat detail kamera berdasarkan ID
      parameters:
      - description: Camera ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Get Camera By ID
      tags:
      - User - Camera
...
  /users/orders:
    get:
      description: Mendapatkan seluruh order milik user yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Get My Orders
      tags:
      - User - Order
    post:
      consumes:
      - application/json
      description: User membuat order rental dengan memilih camera
      parameters:
      - description: Rental Order Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/p2-ip-kartikajls_internal_dto.CreateRentalOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Create Rental Order
      tags:
      - User - Order
  /users/orders/{id}:
    get:
      consumes:
      - application/json
      description: User melihat detail order
      parameters:
      - description: Rental Order ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/p2-ip-kartikajls_internal_helper.Response'
      security:
      - BearerAuth: []
      summary: Get Order By ID
      tags:
      - User - Order
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"